преобразованипе int в double

+++ src/main/java/com/undina/parser_func/service/Function.java
@@ -3,5 +3,5 @@ package com.undina.parser_func.service;
 import java.util.List;
 
 public interface Function {
-    int apply(List<Integer> args);
+    double apply(List<Double> args);
 }
